// types.ts
export interface ServiceCardProps {
  title: string;
  image: string;
  badge: string;
  badgeColor: string;
  estimate: string;
  price: string;
  icon: string;
}

export interface FeatureProps {
  title: string;
  description: string;
  icon: string;
  iconBg: string;
  iconColor: string;
}

export interface ProblemProps {
  title: string;
  description: string;
  icon: string;
  iconBg: string;
  iconColor: string;
}